requirements.txt
biopython==1.74
Django==2.2.1
django-mysql==3.0.0.post1
mysqlclient==1.4.2.post1
numpy==1.16.4
pandas==0.24.2
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
